+++ dbt/include/databricks/macros/relations/materialized_view/create.sql
@@ -1,4 +1,8 @@
 {% macro databricks__get_create_materialized_view_as_sql(relation, sql) -%}
+  {# Column masks are supported in DBSQL, but not yet wired up to the adapter. Return a helpful error until supported. #}
+  {% if column_mask_exists() %}
+    {% do exceptions.raise_compiler_error("Column masks are not yet supported for materialized views.") %}
+  {% endif %}
   {%- set materialized_view = adapter.get_config_from_model(config.model) -%}
   {%- set partition_by = materialized_view.config["partition_by"].partition_by -%}
   {%- set tblproperties = materialized_view.config["tblproperties"].tblproperties -%}